requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
# AI Service Dependencies with compatible versions
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
anthropic==0.7.0
httpx==0.25.2
# Speech processing dependencies
speechrecognition==3.10.0
pydub==0.25.1
# Additional utilities
redis==5.0.1